The second collection view only displays favorited products.
Add notification when a ProductCollectionViewModel is changed.
Every time a product is added or removed the view model
publishes a notification.

The controller reloads the data every time the collection
changes.

Every time a product is favorited the products gets added/removed
to the favorited products collection and UI is automatically reload.