Fixes #8
Products are no longer fetched every time the main view controller
is displayed, only the first time. Both the product collection
and favorite collection view models are listened for rest
notifications, once a notification is trigered an internal
flag in the main view model is marked a loaded. This flag is used
to decide if it is necesary to actually load the collections.

This is not the best solution because if at least one of the request
is successful the flag is set to YES, it does not care if the other
one fails. But for this use case is more than enough.

A better solution will wait for both of the request to be succesful
and only then set the flag to YES.